Thanks, here's a new patch that includes the hostname in the rsync-is-disabled log. I thought about doing that initially but then somehow I missed it.
Darin

On Tue, Jan 5, 2010 at 2:50 PM, John Admanski <jadmanski@google.com> wrote:

> Looks good overall.
>
> On the rsync-is-disabled warning log, I would include the hostname in the
> message.
>
> -- John
>
> On Tue, Jan 5, 2010 at 2:02 PM, Darin Petkov <petkov@google.com> wrote:
>
>> Agreed.
>>
>> Attached is an updated patch.
>>
>> Darin
>>
>>
>> On Tue, Jan 5, 2010 at 12:50 PM, John Admanski <jadmanski@google.com>wrote:
>>
>>> I like less invasive changes, but only up to a point. I just think having
>>> the command generator "turn off" rsync by changing the command to false is a
>>> very un-obvious; I had to think about it for a while before I could convince
>>> myself it was doing the right thing, and I already understand how the copy
>>> method in abstract_ssh.py work.
>>>
>>> -- John
>>>
>>>
>>> On Tue, Jan 5, 2010 at 11:13 AM, Darin Petkov <petkov@google.com> wrote:
>>>
>>>> I thought less invasive changes would be appreciated. No problem, I'll
>>>> fix so that we don't try rsync at all and resend the patch.
>>>>
>>>> Darin
>>>>
>>>>
>>>> On Tue, Jan 5, 2010 at 11:05 AM, John Admanski <jadmanski@google.com>wrote:
>>>>
>>>>> I think disabling rsync by having the make command method return
>>>>> "false" a bit of hack, though. I think it would be cleaner to just make the
>>>>> rsync attempt conditional somehow.
>>>>>
>>>>> On Tue, Jan 5, 2010 at 10:35 AM, Darin Petkov <petkov@google.com>wrote:
>>>>>
>>>>>> Check if rsync is executable on each newly instantiated remote SSH
>>>>>> host.
>>>>>> If it's not, rsync is disabled for future file transfers and scp is
>>>>>> used instead.
>>>>>> If it is, warning messages are logged whenever an rsync-based transfer
>>>>>> fails.
>>>>>>
>>>>>>
>>>>>> _______________________________________________
>>>>>> Autotest mailing list
>>>>>> Autotest@test.kernel.org
>>>>>> http://test.kernel.org/cgi-bin/mailman/listinfo/autotest
>>>>>>
>>>>>>
>>>>>
>>>>
>>>
>>
>
>From 3dca5891ae7da5934a7679f1f3135fe145f2acd5 Mon Sep 17 00:00:00 2001
From: Darin Petkov <petkov@chromium.org>
Date: Tue, 5 Jan 2010 10:24:10 -0800
Subject: [PATCH] Disable rsync if it's not available on the remote host.

Check if rsync is executable on each newly instantiated remote SSH host.
If it's not, rsync is disabled for future file transfers and scp is used
instead.
If it is, warning messages are logged whenever an rsync-based transfer fails.

+++ server/hosts/abstract_ssh.py
@@ -36,6 +36,24 @@ class AbstractSSHHost(SiteHost):
         self.port = port
         self.password = password
 
+        # Check if rsync is available on the remote host. If it's not,
+        # don't try to use it for any future file transfers.
+        self.use_rsync = self._check_rsync()
+        if not self.use_rsync:
+            logging.warn("rsync not available on remote host %s -- disabled",
+                         self.hostname)
+
+
+    def _check_rsync(self):
+        """
+        Check if rsync is available on the remote host.
+        """
+        try:
+            self.run("rsync --version", stdout_tee=None, stderr_tee=None)
+        except error.AutoservRunError:
+            return False
+        return True
+
 
     def _encode_remote_paths(self, paths, escape=True):
         """
@@ -204,13 +222,20 @@ class AbstractSSHHost(SiteHost):
             source = [source]
         dest = os.path.abspath(dest)
 
-        try:
-            remote_source = self._encode_remote_paths(source)
-            local_dest = utils.sh_escape(dest)
-            rsync = self._make_rsync_cmd([remote_source], local_dest,
-                                         delete_dest, preserve_symlinks)
-            utils.run(rsync)
-        except error.CmdError, e:
+        # If rsync is disabled or fails, try scp.
+        try_scp = not self.use_rsync
+        if self.use_rsync:
+            try:
+                remote_source = self._encode_remote_paths(source)
+                local_dest = utils.sh_escape(dest)
+                rsync = self._make_rsync_cmd([remote_source], local_dest,
+                                             delete_dest, preserve_symlinks)
+                utils.run(rsync)
+            except error.CmdError, e:
+                logging.warn("trying scp, rsync failed: %s" % e)
+                try_scp = True
+
+        if try_scp:
             # scp has no equivalent to --delete, just drop the entire dest dir
             if delete_dest and os.path.isdir(dest):
                 shutil.rmtree(dest)
@@ -269,12 +294,19 @@ class AbstractSSHHost(SiteHost):
             source = [source]
         remote_dest = self._encode_remote_paths([dest])
 
-        try:
-            local_sources = [utils.sh_escape(path) for path in source]
-            rsync = self._make_rsync_cmd(local_sources, remote_dest,
-                                         delete_dest, preserve_symlinks)
-            utils.run(rsync)
-        except error.CmdError, e:
+        # If rsync is disabled or fails, try scp.
+        try_scp = not self.use_rsync
+        if self.use_rsync:
+            try:
+                local_sources = [utils.sh_escape(path) for path in source]
+                rsync = self._make_rsync_cmd(local_sources, remote_dest,
+                                             delete_dest, preserve_symlinks)
+                utils.run(rsync)
+            except error.CmdError, e:
+                logging.warn("trying scp, rsync failed: %s" % e)
+                try_scp = True
+
+        if try_scp:
             # scp has no equivalent to --delete, just drop the entire dest dir
             if delete_dest:
                 is_dir = self.run("ls -d %s/" % dest,
